Use position_ instead of position_item
The items in the position_list are positions.
This change makes this easier to understand.
Since we've been dealing with position and position_list in the code block
above not introducing a new variable name reduces cognitive load.

The trailing underscore prevents potential conflict with the position parameter
of the function.