Shorten material_diameter related variables
Shorten old_approximate_material_diameter to:
old_material_diameter
and
new_approximate_material_diameter to
new_material_diameter

The fact that the diameters are approximate can be inferred from the
functions that set them. It is also not particularly relevant in the
context.

The code is now easier to read and understand.
Code lines are now shorter and less of a problem in an editor.