3mf project files saved with an abstract printer would crash on loading when selecting a non abstract printer to load the project with.
This was because the metadata "is_abstract_machine" was being loaded into the non abstract machine. This caused a crash in MachineListModel.py by trying to delete this non abstract machine from a list where it didn't exist.

The solution is to ignore the "is_abstract_machine" metadata when loading settings from saved machines in 3mf files.

CURA-9711