Expose per-model retraction settings.
Related engine PR made it possible to do this. No PR was made to expose these to the front-end as far as I can see, so (since this was a PRwe where very behind on -- like 10 months or so) I choose to do it myself instead of asking the original author.

part of CURA-9876