Specify to build installer per-OS.
Exclusion happens _before_ additional includes unfortunaltely, so the matrix is back to being a plain list if OS's. This also means that, if we don't want to build each OS x OS_id, we have to put that information somewhere else. OS_id didn't really belong in the matrix anyway, since it's a derivative value. It's now replaced with a named environment variable.

The way to exclude and select the 'name' used in the filename for each OS seems a little cumbersome, but I didn't find any better way so far.

(should) implement the 'easy', 'lite' version  of CURA-9880