Attempt to set file associations and fire-wall exceptions.
The fire-wall exceptions work, but the device needs to be restarted. Is that worth it?

A more important problem is that windows overhauled the file associations a few years ago. The approach here doesn't work (but is at least consistent with how it worked before, so it might work for Windows < 10 or unupdated). The current go-to approach is the one used by firefox, and it happens after the installer, at first start-up of the program.

part of CURA-9157