Removed old 4.x settings
The settings:
 - travel_compensate_overlapping_walls_0_enabled
 - fill_perimeter_gaps
 - filter_out_tiny_gaps

Are no longer used with the new Arachne engine. That these settings were
still set in the blocks_base definition is an indication that the current
profiles weren't actually tested with the new Arachne engine.

Please consider doing some actual quality tests with these profiles on your
printers. This is for the benefit of your users experience.

Contributes to CURA-10101