allow bridge speeds to be lower then cool_min_speed
With PP-71 the cool_min_speed is increased to a high values in case the lift_head is enabled. This causes the bridge_speeds to be lower then the cool_min_speed. Which is no problem, so the limit can be removed.
PP-71 / CURA-10211

(cherry picked from commit 6d0be0191340a3d4888662812ee1033305c1196b)