Try before the beta. After it's too late.
A lot was changed in support. The short of it is that we now think towers should be on by default, but aren't _completely_ certain about it. It makes more sense to do make this change before the beta rather than just before stable, so we have a chance to revert should it prove disastrous.