Fix gcode flavors in profiles
For some printer definitions the `machine_gcode_flavor` was entered incorrectly (the display name was used rather than the flavor-key). Previously in combo boxes the flavour would default to the first key (which is "marlin"). Due to recent changes we will not default anymore to the first option, but instead show nothing. By changing the flavor to the correct key the flavors will show up correctly again.

CURA-10374