Align main-based tree-2.0 branch with changes from 'xmas-alpha'.
This is the last change that was missing, originally part of a way larger commit, which is why it wasn't reaplied.

the original commit was part of CURA-9879, this 'manual cherry-pick' was done because there was one branch for CURA-10415 that was based on the 'xmas build' we shated with the public. However, in order to get that out in time, we made some temporary changes, that apparently should not be merged to main. Another branch was made for CURA-10415 that's based on main, but it appeared that this particular change was still missing.