Despite the name, 'Maximum' Fan Speed should be able to go below minimum.
This is because the maximum fan speed is really just the fan speed for layers at or below a certain layer time threshold. Interpolation-logic in the backend (engine) still holds up, even if the values are 'switched' in the sense that the supposed 'max' is lower than the 'min'.

Change was needed since some UM materials had a Maximum Fan Speed set to 15, even though the normal speed and the Minimum Fan Speed are set to 50 in the machine. Despite the confusing naming, from context this seems intentional.

Will fix #15233