Increase min value warning for max extr dev
500 is rather low (100x as low as UltiMaker printers by default = 50.000).
This could result in people (with a Custom FFF printer) to set that value to
unrealistic values such as 5000 ( see project file in #15152 ).
Increasing the warning threshold would still allow people to go below it if
they deem it necessary.

Giving a warning at 10.000 (corresponds to 125 micrometer length with a linewidth of 0.4 mm)
seems reasonable imo

Contributes to CURA-10500