Use SHA of the commit instead of no_commits since last tag
We determine the version of a conan package based on a Git tag and then count the number of commits between the last tag
and the current tag to get the +testing_6 or +testing_13 after the alpha/beta designation etc.
This will give us logical sequential numbers with which we can determine which version was later.
But the method can't take into account tags created on release branches, a branch parallel to the main branch, once
these are merged, after a tag then suddenly the first tag it encounters is closer then it was before, this means that we
then accidentally create a second package with the same version and this starts tripping things up.
 Now these logical statements aren't necessary to determine the latest version, according to semver everything after the
  + is ignored and both versions should be compatible.
An easy fix is to replace the testing_<no_of_commits_since_last_tag> with testing_<commit_hash> that way we don't have
two conan packages created from different commits anymore, which is the most common root cause of our problems in this
flow.