Disable Small Skin Area feature for now :-/
We're not sure why this introduces a boatload of small segments into the infill, and we're out of time to try and fix this before the beta. Re-introducing after the beta seems like a bad idea, so this'll have to wait until 5.5 unfortunately.

Done as part of fixing for CURA-10500, but for the original Small Skin Area feature, see CURA-10201. For the ticket to (fix? and) reintroduce the feature, see CURA-10670.