Update BridgeTemperatureAdjustment.py
Remove code that added the PP name to the gcode.  Left messages.

Update BridgeTemperatureAdjustment.py

Fixed a conflict if AddCoolingProfile runs first and is set to "By Feature".

Create BridgeTemperatureAdjustment.py

A post to alter the Hot End temperature when printing bridges.  This is a response to a request on GitHub.