Reduce maximum warning value for small top/bottom skin.
While the current solution works well for values smaller than this (and that's mostly what we'd need it for), it becomes clear beyond this value (if you'd like to set it that hight for some reason) that the current solution needs to be improved (for example by adding these small skin areas to the already existing wall areas. -- Lift the solution out of the fill as it where.

WHat happens when it becomes larger it that it's very likely that it'll generate a dot as middle walls on the 'outsides' of any small area connected to a larger one.

done as part of CURA-10670