Refactor object reload to handle 3mf files
Refactor object reload to better handling of 3mf files

Make a single file read for all objects from that file, and then
replace the objects in scene to the new ones, matching by name.

This is also required for an Uranium update to handle the automatic
file reload when a file change is detected.

3mf Files can have multiple objects in them. When that is the case,
and the file is updated, all objects from that file where being
replaced by the first object in the file.