Bundling the Gradualflow plugin
This commit simplifies the deploy method in conanfile.py by reducing the number of lines of code and refactor the
handling of the Gradualflow plugin in the deploy method. In conandata.yml, the paths for the Gradualflow plugin
resources have been adjusted for correct deployment. This update aims to fix the pyinstaller deployment failures
associated with the Gradualflow plugin. It resolves the CURA-10951 tracking issue.

Contributes CURA-10951