Simplified All Installer workflow
No need to make exceptions in this workflow for different OSes
it's called the **All** Installer for a reason.

If people need a specific OS installer run the linux, macos or windows installer itself.

Added a time-out to the macos workflow if the self-hosted isn't online.
Moved the ARM64 jobs to the end and added an `always()` condition to the update of the release, such that the
GH hosted runners will at least update.

Add missing download for MacOS X64

Contributes to CURA-10475 and CURA-10591