Only filter files in directories starting with .
Mac builds were missing some definitions, which happened because there was
a . in the directory. The reason why we filtered these in the first place
was because Apple doesn't notarize/sign an App correctly if the directory
starts with a . e.q.: .dylibs in scipy and shapely.

Made the filtering less aggressive

Contributes to CURA-11014