Update PyQt6 dependencies to latest versions
PyQt6 and PyQt6-sip along with their dependencies are updated to newer versions (6.5.2 and 13.5.2) in both requirements.txt and Uranium/requirements.txt. This is to ensure the application uses the latest and most secure versions of the libraries, potentially improving application performance and security. These updates also ensure compatibility with current and upcoming features and bug fixes.

Contributes to CURA-11137