Fix translations.
Given that we're still in negotiation with the new translation service after our trial period ran out, we've sadly had to rely on machine translations for these. Czech was missing the new Gradual Flow settings, and German had an unrelated setting description somehow end up as a label for another.

a late part of CURA-10994