Modernize conanfile.py
This commit updates the Uranium dependency in the Conan file and introduces an internationalization (i18n) option.
It refactor the devtool checks to check for the "enable_i18n" option instead.
Lastly, cleanups and adjustments are done in preparation for enabling i18n in the build process.

Contributes to CURA-10831