Add native CAD plugin to Enterprise edition
The native CAD plugin has been added to the requirements and reference copy operations in the Conan data file and Conan file script.
Now the native CAD plugin will only be included if the enterprise build flag is set. This ensures that unnecessary dependencies are not brought into non-enterprise builds.