Use the activeIntent from intentmanager instead of active stack
Otherwise the behavior of the model changes depending on what stack you
have active. I think that this was caused by some code duplication where one
copy of the code was updated but not the other one. This bug was here for quite
some time, but strangely enough, it was never found...

CURA-11854