Fix RecentFilesMenu.qml onObjectRemoved incompatible arguments
"share/cura/resources/qml/Menus/RecentFilesMenu.qml:39:
TypeError: Passing incompatible arguments to C++ functions from JavaScript is not allowed."

This was passing the Instantiator index, which is a Number to menu.removeItem
which is expecting an object.  Add the missing index argument.  I
found Qt 5.7 had two arguments so it has been there for some time.