Update PurgeLinesAndUnload.py
Change the comment line location from the end of data[1] to the first line of the "moves".

Update PurgeLinesAndUnload.py

Move the "start from" comment from the end of data[1] to the first line of the "moves" list.

Update PurgeLinesAndUnload.py

Change the comment line location from the end of data[1] to the first line of the "moves".