StartSliceJob.py - another different attempt
This time, I'm not trying to correct the regex, as it seems impossible without using recursion to ensure that all brackets are closed. Instead, I manually iterate over the `expression` capture group, and if it is misaligned in regards to brackets, continue reading into the `extruder_nr_expr` group until I find a comma outside of any brackets, and then modify the two fields accordingly (moving the same amount of data out of one and into the other) before sending them on to the expression parser.