Fix "Open in Cura" button on Thingiverse
Make the "Open in Cura" button on Thingiverse work by informing
the system that Cura is the application to handle URLs with the
cura scheme, and that it can handle URLs, not just local files.

This fixes this error message visible in Firefox' console when clicking
the button on the website:
Prevented navigation to “cura://open?file=https%3A%2F%2Fwww.thingiverse.com%2Fdownload%3A14134689” due to an unknown protocol.

See https://github.com/bambulab/BambuStudio/commit/62006bba30f39